Factor out ConscepterTest code
